Change ~/${PWD/#$HOME} to ~/pw/jobs/script_submitter/tmp
There are two issues with ~/${PWD/#$HOME}:
1. it uses Bash search/replace with anchors which most users do
   not recognize
2. it assumes that the PW workspace username is the same as the
   cluster username

Solution: switch to a different directory. The difficulty in
a workflow framework is that generating a unique workflow
directory (i.e. a uuid) is hard to pass from one job step to
another. So as an initial test, choose a specific tmp directory
to check if this change addresses issue #2, above.